pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "explainable-ai-tutor"
version = "0.1.0"
description = "AI-powered educational explainer for PowerPoint presentations using fine-tuned Mistral 7B"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.4.0",
    "transformers==4.49.0",
    "peft==0.17.1",
    "datasets==4.3.0",
    "gradio==5.49.1",
    "python-pptx==1.0.0",
    "Pillow==11.2.1",
    "accelerate",
    "bitsandbytes",
    "sentencepiece",
    "protobuf",
    "trl==0.15.2",
    "evaluate>=0.4.6",
    "sacrebleu>=2.5.0",
    "rouge-score>=0.1.2",
    "tqdm",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.5",
    "mypy>=1.10",
    "jupyter",
    "ipykernel",
    "ipywidgets",
    "nbformat>=5.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
ignore_missing_imports = true

